@@ -4,22 +4,22 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   drift_local_storage_inspector:
     path: ../
-  storage_inspector: ^0.3.0
-  drift: ^2.2.0
+  storage_inspector: ^0.4.0
+  drift: ^2.11.1
 
 dev_dependencies:
-  build_runner: ^2.3.2
-  drift_dev: ^2.2.0+1
+  build_runner: ^2.4.6
+  drift_dev: ^2.11.1
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^2.0.3
 
 flutter:
   uses-material-design: true
